@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import {Link} from 'next-url-prettifier'
 import {Router} from '../routes'
 
@@ -29,6 +30,6 @@ export default class GreetingPage extends React.Component {
 }
 
 GreetingPage.propTypes = {
-  lang: React.PropTypes.string,
-  name: React.PropTypes.string
+  lang: PropTypes.string,
+  name: PropTypes.string
 }
