@@ -1,9 +1,10 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import sideEffect from './side-effect'
 
 class Head extends React.Component {
   static contextTypes = {
-    headManager: React.PropTypes.object
+    headManager: PropTypes.object
   }
 
   render () {
