@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { AppContainer } from 'react-hot-loader'
 import shallowEquals from './shallow-equals'
 import { warn } from './utils'
