@@ -1,5 +1,6 @@
 import { resolve, format, parse } from 'url'
-import React, { Component, Children, PropTypes } from 'react'
+import React, { Component, Children } from 'react'
+import PropTypes from 'prop-types'
 import Router from './router'
 import { warn, execOnce, getLocationOrigin } from './utils'
 
